speaker making low noise off manual set

+++ project3/project3/audio.h
@@ -5,6 +5,6 @@
 
 void audio_on(void);
 void audio_off(void);
-void play_freq(uint16_t freq, uint16_t volume);
+void play_freq(uint16_t freq, uint8_t volume);
 
 #endif
